requirements.txt
ultralytics
numpy>=1.18.5
opencv-python>=4.5.4.60
Pillow>=7.1.2
filterpy==1.4.5
scikit-image==0.19.3
Image

# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

pynput
pywin32
